+++ src/libshadertrap/src/command_set_uniform.cc
@@ -22,7 +22,7 @@ namespace shadertrap {
 
 CommandSetUniform::CommandSetUniform(std::unique_ptr<Token> start_token,
                                      std::string program_identifier,
-                                     uint32_t location, UniformValue value)
+                                     size_t location, UniformValue value)
     : Command(std::move(start_token)),
       program_identifier_(std::move(program_identifier)),
       location_(location),
